Move routes to v1
