+++ src/StoreApi/Routes/V1/Cart.php
@@ -1,19 +1,19 @@
 <?php
-namespace Automattic\WooCommerce\Blocks\StoreApi\Routes;
+namespace Automattic\WooCommerce\Blocks\StoreApi\Routes\V1;
 
 /**
- * ProductTags class.
+ * Cart class.
  *
  * @internal This API is used internally by Blocks--it is still in flux and may be subject to revisions.
  */
-class ProductTags extends AbstractTermsRoute {
+class Cart extends AbstractCartRoute {
 	/**
 	 * Get the path of this REST route.
 	 *
 	 * @return string
 	 */
 	public function get_path() {
-		return '/products/tags';
+		return '/cart';
 	}
 
 	/**
@@ -27,20 +27,22 @@ class ProductTags extends AbstractTermsRoute {
 				'methods'             => \WP_REST_Server::READABLE,
 				'callback'            => [ $this, 'get_response' ],
 				'permission_callback' => '__return_true',
-				'args'                => $this->get_collection_params(),
+				'args'                => [
+					'context' => $this->get_context_param( [ 'default' => 'view' ] ),
+				],
 			],
 			'schema' => [ $this->schema, 'get_public_item_schema' ],
 		];
 	}
 
 	/**
-	 * Get a collection of terms.
+	 * Handle the request and return a valid response for this endpoint.
 	 *
 	 * @throws RouteException On error.
 	 * @param \WP_REST_Request $request Request object.
 	 * @return \WP_REST_Response
 	 */
 	protected function get_route_response( \WP_REST_Request $request ) {
-		return $this->get_terms_response( 'product_tag', $request );
+		return rest_ensure_response( $this->schema->get_item_response( $this->cart_controller->get_cart_instance() ) );
 	}
 }
