+++ src/StoreApi/Routes/V1/CartRemoveItem.php
@@ -1,19 +1,23 @@
 <?php
-namespace Automattic\WooCommerce\Blocks\StoreApi\Routes;
+namespace Automattic\WooCommerce\Blocks\StoreApi\Routes\V1;
+
+use Automattic\WooCommerce\Blocks\StoreApi\Utilities\DraftOrderTrait;
 
 /**
- * CartUpdateItem class.
+ * CartRemoveItem class.
  *
  * @internal This API is used internally by Blocks--it is still in flux and may be subject to revisions.
  */
-class CartUpdateItem extends AbstractCartRoute {
+class CartRemoveItem extends AbstractCartRoute {
+	use DraftOrderTrait;
+
 	/**
 	 * Get the path of this REST route.
 	 *
 	 * @return string
 	 */
 	public function get_path() {
-		return '/cart/update-item';
+		return '/cart/remove-item';
 	}
 
 	/**
@@ -28,13 +32,9 @@ class CartUpdateItem extends AbstractCartRoute {
 				'callback'            => [ $this, 'get_response' ],
 				'permission_callback' => '__return_true',
 				'args'                => [
-					'key'      => [
-						'description' => __( 'Unique identifier (key) for the cart item to update.', 'woo-gutenberg-products-block' ),
+					'key' => [
+						'description' => __( 'Unique identifier (key) for the cart item.', 'woo-gutenberg-products-block' ),
 						'type'        => 'string',
-					],
-					'quantity' => [
-						'description' => __( 'New quantity of the item in the cart.', 'woo-gutenberg-products-block' ),
-						'type'        => 'integer',
 					],
 				],
 			],
@@ -51,12 +51,31 @@ class CartUpdateItem extends AbstractCartRoute {
 	 * @return \WP_REST_Response
 	 */
 	protected function get_route_post_response( \WP_REST_Request $request ) {
-		$cart = $this->cart_controller->get_cart_instance();
+		$cart      = $this->cart_controller->get_cart_instance();
+		$cart_item = $this->cart_controller->get_cart_item( $request['key'] );
 
-		if ( isset( $request['quantity'] ) ) {
-			$this->cart_controller->set_cart_item_quantity( $request['key'], $request['quantity'] );
+		if ( empty( $cart_item ) ) {
+			throw new RouteException( 'woocommerce_rest_cart_invalid_key', __( 'Cart item no longer exists or is invalid.', 'woo-gutenberg-products-block' ), 409 );
 		}
 
+		$cart->remove_cart_item( $request['key'] );
+		$this->maybe_release_stock();
+
 		return rest_ensure_response( $this->schema->get_item_response( $cart ) );
+	}
+
+	/**
+	 * If there is a draft order, releases stock.
+	 *
+	 * @return void
+	 */
+	protected function maybe_release_stock() {
+		$draft_order_id = $this->get_draft_order_id();
+
+		if ( ! $draft_order_id ) {
+			return;
+		}
+
+		wc_release_stock_for_order( $draft_order_id );
 	}
 }
