+++ src/StoreApi/Routes/V1/CartUpdateItem.php
@@ -1,23 +1,19 @@
 <?php
-namespace Automattic\WooCommerce\Blocks\StoreApi\Routes;
-
-use Automattic\WooCommerce\Blocks\StoreApi\Utilities\DraftOrderTrait;
+namespace Automattic\WooCommerce\Blocks\StoreApi\Routes\V1;
 
 /**
- * CartRemoveItem class.
+ * CartUpdateItem class.
  *
  * @internal This API is used internally by Blocks--it is still in flux and may be subject to revisions.
  */
-class CartRemoveItem extends AbstractCartRoute {
-	use DraftOrderTrait;
-
+class CartUpdateItem extends AbstractCartRoute {
 	/**
 	 * Get the path of this REST route.
 	 *
 	 * @return string
 	 */
 	public function get_path() {
-		return '/cart/remove-item';
+		return '/cart/update-item';
 	}
 
 	/**
@@ -32,9 +28,13 @@ class CartRemoveItem extends AbstractCartRoute {
 				'callback'            => [ $this, 'get_response' ],
 				'permission_callback' => '__return_true',
 				'args'                => [
-					'key' => [
-						'description' => __( 'Unique identifier (key) for the cart item.', 'woo-gutenberg-products-block' ),
+					'key'      => [
+						'description' => __( 'Unique identifier (key) for the cart item to update.', 'woo-gutenberg-products-block' ),
 						'type'        => 'string',
+					],
+					'quantity' => [
+						'description' => __( 'New quantity of the item in the cart.', 'woo-gutenberg-products-block' ),
+						'type'        => 'integer',
 					],
 				],
 			],
@@ -51,31 +51,12 @@ class CartRemoveItem extends AbstractCartRoute {
 	 * @return \WP_REST_Response
 	 */
 	protected function get_route_post_response( \WP_REST_Request $request ) {
-		$cart      = $this->cart_controller->get_cart_instance();
-		$cart_item = $this->cart_controller->get_cart_item( $request['key'] );
+		$cart = $this->cart_controller->get_cart_instance();
 
-		if ( empty( $cart_item ) ) {
-			throw new RouteException( 'woocommerce_rest_cart_invalid_key', __( 'Cart item no longer exists or is invalid.', 'woo-gutenberg-products-block' ), 409 );
+		if ( isset( $request['quantity'] ) ) {
+			$this->cart_controller->set_cart_item_quantity( $request['key'], $request['quantity'] );
 		}
-
-		$cart->remove_cart_item( $request['key'] );
-		$this->maybe_release_stock();
 
 		return rest_ensure_response( $this->schema->get_item_response( $cart ) );
-	}
-
-	/**
-	 * If there is a draft order, releases stock.
-	 *
-	 * @return void
-	 */
-	protected function maybe_release_stock() {
-		$draft_order_id = $this->get_draft_order_id();
-
-		if ( ! $draft_order_id ) {
-			return;
-		}
-
-		wc_release_stock_for_order( $draft_order_id );
 	}
 }
