+++ src/StoreApi/Routes/V1/ProductTags.php
@@ -1,19 +1,19 @@
 <?php
-namespace Automattic\WooCommerce\Blocks\StoreApi\Routes;
+namespace Automattic\WooCommerce\Blocks\StoreApi\Routes\V1;
 
 /**
- * ProductCategories class.
+ * ProductTags class.
  *
  * @internal This API is used internally by Blocks--it is still in flux and may be subject to revisions.
  */
-class ProductCategories extends AbstractTermsRoute {
+class ProductTags extends AbstractTermsRoute {
 	/**
 	 * Get the path of this REST route.
 	 *
 	 * @return string
 	 */
 	public function get_path() {
-		return '/products/categories';
+		return '/products/tags';
 	}
 
 	/**
@@ -41,6 +41,6 @@ class ProductCategories extends AbstractTermsRoute {
 	 * @return \WP_REST_Response
 	 */
 	protected function get_route_response( \WP_REST_Request $request ) {
-		return $this->get_terms_response( 'product_cat', $request );
+		return $this->get_terms_response( 'product_tag', $request );
 	}
 }
