+++ src/deprecated.php
@@ -5,6 +5,7 @@
  * This file is autoloaded via composer.json and maps the old namespaces to new namespaces.
  */
 
+// Unversioned schema to v1 schema.
 class_alias( Automattic\WooCommerce\Blocks\StoreApi\Schemas\V1\AbstractAddressSchema::class, Automattic\WooCommerce\Blocks\StoreApi\Schemas\AbstractAddressSchema::class );
 class_alias( Automattic\WooCommerce\Blocks\StoreApi\Schemas\V1\AbstractSchema::class, Automattic\WooCommerce\Blocks\StoreApi\Schemas\AbstractSchema::class );
 class_alias( Automattic\WooCommerce\Blocks\StoreApi\Schemas\V1\BillingAddressSchema::class, Automattic\WooCommerce\Blocks\StoreApi\Schemas\BillingAddressSchema::class );
@@ -25,3 +26,33 @@ class_alias( Automattic\WooCommerce\Blocks\StoreApi\Schemas\V1\ProductReviewSche
 class_alias( Automattic\WooCommerce\Blocks\StoreApi\Schemas\V1\ProductSchema::class, Automattic\WooCommerce\Blocks\StoreApi\Schemas\ProductSchema::class );
 class_alias( Automattic\WooCommerce\Blocks\StoreApi\Schemas\V1\ShippingAddressSchema::class, Automattic\WooCommerce\Blocks\StoreApi\Schemas\ShippingAddressSchema::class );
 class_alias( Automattic\WooCommerce\Blocks\StoreApi\Schemas\V1\TermSchema::class, Automattic\WooCommerce\Blocks\StoreApi\Schemas\TermSchema::class );
+
+// Unversioned routes to v1 routes.
+class_alias( Automattic\WooCommerce\Blocks\StoreApi\Routes\V1\AbstractCartRoute::class, Automattic\WooCommerce\Blocks\StoreApi\Routes\AbstractCartRoute::class );
+class_alias( Automattic\WooCommerce\Blocks\StoreApi\Routes\V1\AbstractRoute::class, Automattic\WooCommerce\Blocks\StoreApi\Routes\AbstractRoute::class );
+class_alias( Automattic\WooCommerce\Blocks\StoreApi\Routes\V1\AbstractTermsRoute::class, Automattic\WooCommerce\Blocks\StoreApi\Routes\AbstractTermsRoute::class );
+class_alias( Automattic\WooCommerce\Blocks\StoreApi\Routes\V1\Batch::class, Automattic\WooCommerce\Blocks\StoreApi\Routes\Batch::class );
+class_alias( Automattic\WooCommerce\Blocks\StoreApi\Routes\V1\Cart::class, Automattic\WooCommerce\Blocks\StoreApi\Routes\Cart::class );
+class_alias( Automattic\WooCommerce\Blocks\StoreApi\Routes\V1\CartAddItem::class, Automattic\WooCommerce\Blocks\StoreApi\Routes\CartAddItem::class );
+class_alias( Automattic\WooCommerce\Blocks\StoreApi\Routes\V1\CartApplyCoupon::class, Automattic\WooCommerce\Blocks\StoreApi\Routes\CartApplyCoupon::class );
+class_alias( Automattic\WooCommerce\Blocks\StoreApi\Routes\V1\CartCoupons::class, Automattic\WooCommerce\Blocks\StoreApi\Routes\CartCoupons::class );
+class_alias( Automattic\WooCommerce\Blocks\StoreApi\Routes\V1\CartCouponsByCode::class, Automattic\WooCommerce\Blocks\StoreApi\Routes\CartCouponsByCode::class );
+class_alias( Automattic\WooCommerce\Blocks\StoreApi\Routes\V1\CartExtensions::class, Automattic\WooCommerce\Blocks\StoreApi\Routes\CartExtensions::class );
+class_alias( Automattic\WooCommerce\Blocks\StoreApi\Routes\V1\CartItems::class, Automattic\WooCommerce\Blocks\StoreApi\Routes\CartItems::class );
+class_alias( Automattic\WooCommerce\Blocks\StoreApi\Routes\V1\CartItemsByKey::class, Automattic\WooCommerce\Blocks\StoreApi\Routes\CartItemsByKey::class );
+class_alias( Automattic\WooCommerce\Blocks\StoreApi\Routes\V1\CartRemoveCoupon::class, Automattic\WooCommerce\Blocks\StoreApi\Routes\CartRemoveCoupon::class );
+class_alias( Automattic\WooCommerce\Blocks\StoreApi\Routes\V1\CartRemoveItem::class, Automattic\WooCommerce\Blocks\StoreApi\Routes\CartRemoveItem::class );
+class_alias( Automattic\WooCommerce\Blocks\StoreApi\Routes\V1\CartSelectShippingRate::class, Automattic\WooCommerce\Blocks\StoreApi\Routes\CartSelectShippingRate::class );
+class_alias( Automattic\WooCommerce\Blocks\StoreApi\Routes\V1\CartUpdateCustomer::class, Automattic\WooCommerce\Blocks\StoreApi\Routes\CartUpdateCustomer::class );
+class_alias( Automattic\WooCommerce\Blocks\StoreApi\Routes\V1\CartUpdateItem::class, Automattic\WooCommerce\Blocks\StoreApi\Routes\CartUpdateItem::class );
+class_alias( Automattic\WooCommerce\Blocks\StoreApi\Routes\V1\Checkout::class, Automattic\WooCommerce\Blocks\StoreApi\Routes\Checkout::class );
+class_alias( Automattic\WooCommerce\Blocks\StoreApi\Routes\V1\ProductAttributes::class, Automattic\WooCommerce\Blocks\StoreApi\Routes\ProductAttributes::class );
+class_alias( Automattic\WooCommerce\Blocks\StoreApi\Routes\V1\ProductAttributesById::class, Automattic\WooCommerce\Blocks\StoreApi\Routes\ProductAttributesById::class );
+class_alias( Automattic\WooCommerce\Blocks\StoreApi\Routes\V1\ProductAttributeTerms::class, Automattic\WooCommerce\Blocks\StoreApi\Routes\ProductAttributeTerms::class );
+class_alias( Automattic\WooCommerce\Blocks\StoreApi\Routes\V1\ProductCategories::class, Automattic\WooCommerce\Blocks\StoreApi\Routes\ProductCategories::class );
+class_alias( Automattic\WooCommerce\Blocks\StoreApi\Routes\V1\ProductCategoriesById::class, Automattic\WooCommerce\Blocks\StoreApi\Routes\ProductCategoriesById::class );
+class_alias( Automattic\WooCommerce\Blocks\StoreApi\Routes\V1\ProductCollectionData::class, Automattic\WooCommerce\Blocks\StoreApi\Routes\ProductCollectionData::class );
+class_alias( Automattic\WooCommerce\Blocks\StoreApi\Routes\V1\ProductReviews::class, Automattic\WooCommerce\Blocks\StoreApi\Routes\ProductReviews::class );
+class_alias( Automattic\WooCommerce\Blocks\StoreApi\Routes\V1\Products::class, Automattic\WooCommerce\Blocks\StoreApi\Routes\Products::class );
+class_alias( Automattic\WooCommerce\Blocks\StoreApi\Routes\V1\ProductsById::class, Automattic\WooCommerce\Blocks\StoreApi\Routes\ProductsById::class );
+class_alias( Automattic\WooCommerce\Blocks\StoreApi\Routes\V1\ProductTags::class, Automattic\WooCommerce\Blocks\StoreApi\Routes\ProductTags::class );
